refactor(crouton-sales): give the outstanding rule one home before adding a third reader (#1867)
👤 For humans
"Still to deliver" was written out twice — once as TypeScript, once as
hand-rolled SQL inside the Bestellingen counter. Two copies of one sentence
drift, and the per-product view was about to add a third. Now there is one
predicate and both queries call it, so they cannot disagree about what
"still waiting" means.

🤖 For agents
- location-handover.ts: + locationBlocksDeliverySql({ bumpId, requiresHandover }),
  the SQL half of locationBlocksDelivery. NULL requiresHandover must read as
  REQUIRING; SQL three-valued logic won't do that on its own (a bare
  `requiresHandover = 1` is NULL, not false, for a legacy row, silently
  dropping it) — hence the explicit isNull(...) OR eq(..., true).
- orders.get.ts countOutstanding: drops its inline copy of the condition and
  calls the shared predicate. Behaviour-preserving; the emitted WHERE is the
  same. Unused isNull/or imports removed.
- personnel-condition.ts (new): personnelConditionOn(column, value) +
  parsePersonnelMode — the staff-order rule over an INJECTED column.
  personnel-filter.ts becomes the thin binding that supplies salesOrders.
  Needed because that layer import (~~/layers/sales/...) is unresolvable from
  a package unit test and from a query that receives its tables as parameters.
- personnel-filter.ts deliberately does NOT re-export parsePersonnelMode:
  Nitro auto-imports every server/utils/ export, so a re-export makes the
  symbol resolvable from two modules and Nuxt warns "Duplicated imports".
